rag:qrag/requirements.txt
openai>=1.6.0
pinecone-client==2.2.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3 